Move LevelTextureStorage to trview.app
This is so it can be tested.
