+++ trview.app/Graphics/LevelTextureStorage.cpp
@@ -1,4 +1,3 @@
-#include "stdafx.h"
 #include "LevelTextureStorage.h"
 #include "TextureStorage.h"
