+++ trview.app/Graphics/TextureStorage.cpp
@@ -1,4 +1,3 @@
-#include "stdafx.h"
 #include "TextureStorage.h"
 
 namespace trview
