+++ trview.app/trview.app.vcxproj
@@ -36,8 +36,10 @@
     <ClCompile Include="Geometry\TransparentTriangle.cpp" />
     <ClCompile Include="Graphics\ILevelTextureStorage.cpp" />
     <ClCompile Include="Graphics\ITextureStorage.cpp" />
+    <ClCompile Include="Graphics\LevelTextureStorage.cpp" />
     <ClCompile Include="Graphics\SectorHighlight.cpp" />
     <ClCompile Include="Graphics\SelectionRenderer.cpp" />
+    <ClCompile Include="Graphics\TextureStorage.cpp" />
     <ClCompile Include="Menus\AlternateGroupToggler.cpp" />
     <ClCompile Include="Menus\DirectoryListing.cpp" />
     <ClCompile Include="Menus\FileDropper.cpp" />
@@ -93,8 +95,10 @@
     <ClInclude Include="Geometry\Triangle.h" />
     <ClInclude Include="Graphics\ILevelTextureStorage.h" />
     <ClInclude Include="Graphics\ITextureStorage.h" />
+    <ClInclude Include="Graphics\LevelTextureStorage.h" />
     <ClInclude Include="Graphics\SectorHighlight.h" />
     <ClInclude Include="Graphics\SelectionRenderer.h" />
+    <ClInclude Include="Graphics\TextureStorage.h" />
     <ClInclude Include="Menus\AlternateGroupToggler.h" />
     <ClInclude Include="Menus\DirectoryListing.h" />
     <ClInclude Include="Menus\FileDropper.h" />
