+++ trview.app/trview.app.vcxproj.filters
@@ -151,6 +151,12 @@
     <ClCompile Include="Camera\CameraInput.cpp">
       <Filter>Camera</Filter>
     </ClCompile>
+    <ClCompile Include="Graphics\LevelTextureStorage.cpp">
+      <Filter>Graphics</Filter>
+    </ClCompile>
+    <ClCompile Include="Graphics\TextureStorage.cpp">
+      <Filter>Graphics</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Camera\Camera.h">
@@ -320,6 +326,12 @@
     </ClInclude>
     <ClInclude Include="Camera\CameraInput.h">
       <Filter>Camera</Filter>
+    </ClInclude>
+    <ClInclude Include="Graphics\LevelTextureStorage.h">
+      <Filter>Graphics</Filter>
+    </ClInclude>
+    <ClInclude Include="Graphics\TextureStorage.h">
+      <Filter>Graphics</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
