+++ trview/trview.vcxproj
@@ -192,8 +192,6 @@
     <ClInclude Include="StaticMesh.h" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="targetver.h" />
-    <ClInclude Include="LevelTextureStorage.h" />
-    <ClInclude Include="TextureStorage.h" />
     <ClInclude Include="trview.h" />
     <ClInclude Include="Viewer.h" />
   </ItemGroup>
@@ -214,8 +212,6 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="LevelTextureStorage.cpp" />
-    <ClCompile Include="TextureStorage.cpp" />
     <ClCompile Include="trview.cpp" />
     <ClCompile Include="Viewer.cpp" />
   </ItemGroup>
